+++ effect_subject/forms/__init__.py
@@ -19,7 +19,6 @@ from .lp_csf_form import LpCsfForm
 from .medical_history_form import MedicalHistoryForm
 from .medication_adherence_form import MedicationAdherenceForm
 from .mental_status_form import MentalStatusForm
-from .microbiology_form import MicrobiologyForm
 from .patient_history_form import PatientHistoryForm
 from .patient_treatment_day14_form import PatientTreatmentDay14Form
 from .patient_treatment_form import PatientTreatmentForm
@@ -27,4 +26,5 @@ from .signs_and_symptoms_form import SignsAndSymptomsForm
 from .subject_requisition_form import SubjectRequisitionForm
 from .subject_visit_form import SubjectVisitForm
 from .subject_visit_missed_form import SubjectVisitMissedForm
+from .tb_diagnostics_form import TbDiagnosticsForm
 from .vital_signs_form import VitalSignsForm